+++ commands/misc/iam.js
@@ -88,6 +88,9 @@ class Iam extends Command {
         if (member.roles.find(r => r === guildRole.id)) {
             return message.channel.createMessage(':x: You already have this role');
         }
+        if (this.getHighestRole(client.bot.user.id, message.channel.guild) && (guildRole.position > this.getHighestRole(client.bot.user.id, message.channel.guild).position)) {
+            return message.channel.createMessage(`:x: The role \`${guildRole.name}\` is higher than my highest role, therefore, i can't give it to you :c`);
+        }
         await member.addRole(guildRole.id);
         return message.channel.createMessage(":white_check_mark: Alright, i gave you the role `" + guildRole.name + "`");
     }