+++ commands/settings/setfarewells.js
@@ -38,7 +38,7 @@ class SetFarewells extends Command {
             category: 'settings',
             description: 'This command allows you to change the settings of the farewells system',
             usage: '{prefix}setfarewells',
-            externalDoc: 'https://github.com/ParadoxalCorp/Felix-documentation/blob/master/README.md#greetings-and-farewell-system'
+            externalDoc: 'https://github.com/ParadoxalCorp/felix-production/blob/master/usage.md#greetings-and-farewell-system'
         };
         this.conf = {
             requireDB: true,